Рефакторинг
