+++ src/data.js
@@ -1,7 +1,7 @@
-// import {getRandomNumber} from './utils.js';
+import {getRandomNumber} from './utils.js';
 
 // Данные для sorting фильтра
-export const sortFilterArr = [
+export const sortFilterData = [
   {
   	href: `#`,
   	title: `default`
@@ -15,3 +15,45 @@ export const sortFilterArr = [
   	title: `date down`
   }
 ];
+
+export const filterData = [
+  {
+    title: `All`,
+    count: getRandomNumber(0, 50)
+  },
+  {
+    title: `Overdue`,
+    count: getRandomNumber(0, 20)
+  },
+  {
+    title: `Today`,
+    count: getRandomNumber(0, 20)
+  },
+  {
+    title: `Favorites`,
+    count: getRandomNumber(0, 20)
+  },
+  {
+    title: `Repeating`,
+    count: getRandomNumber(0, 20)
+  },
+  {
+    title: `Tags`,
+    count: getRandomNumber(0, 20)
+  },
+  {
+    title: `Archive`,
+    count: getRandomNumber(0, 50)
+  }
+];
+
+// Данные для фильтра арточек
+// export const filterData = [
+//   `All`,
+//   `Overdue`,
+//   `Today`,
+//   `Favorites`,
+//   `Repeating`,
+//   `Tags`,
+//   `Archive`
+// ];
